refactor : 게시판, 내활동에 관리자, 회원 필터 적용 / 관리자 및 회원에 따라 수정, 삭제, 좋아요, 댓글, 투표 버튼 다르게 표시 / 댓글 프사 추가 / 글 작성 필수조건 추가 / 목록으로 버튼 추가 design : 투표 ui 수정 fix : no 파라미터 예외처리 / 내좋아요 페이징 수정

+++ HoneyPot/src/main/java/com/hp/app/board/controller/BoardController.java
@@ -1,8 +1,6 @@
 package com.hp.app.board.controller;
 
 import java.io.File;
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -41,25 +39,27 @@ public class BoardController {
 	@GetMapping("board/free")
 	public String getFreeList(@RequestParam(defaultValue="1") String p, @RequestParam Map<String, String> searchVo, Model model, HttpSession session) {
 		
-		try {
-			
-			searchVo.put("cno", "1");
-			
-			int listCount = service.countBoard(searchVo);
-			int currentPage = Integer.parseInt(p);
-			int pageLimit = 5;
-			int boardLimit = 8;
-			PageVo pv = new PageVo(listCount, currentPage, pageLimit, boardLimit);
-		
-			List<BoardVo> voList = service.getList(pv, searchVo);
-			
-			model.addAttribute("voList", voList);
-			model.addAttribute("pv", pv);
-			model.addAttribute("searchVo", searchVo);
-			
-		}catch (Exception e) {
-			e.printStackTrace();
+		//관리자도, 회원도 아니면 리턴
+		AdminVo loginAdmin = (AdminVo) session.getAttribute("loginAdmin");
+		MemberVo loginMember = (MemberVo) session.getAttribute("loginMember");
+		if(loginAdmin == null && loginMember == null) {
+			session.setAttribute("alertMsg", "로그인 후 이용해주세요.");
+			return "redirect:/member/mlogin";
 		}
+
+		searchVo.put("cno", "1");
+		
+		int listCount = service.countBoard(searchVo);
+		int currentPage = Integer.parseInt(p);
+		int pageLimit = 5;
+		int boardLimit = 8;
+		PageVo pv = new PageVo(listCount, currentPage, pageLimit, boardLimit);
+	
+		List<BoardVo> voList = service.getList(pv, searchVo);
+		
+		model.addAttribute("voList", voList);
+		model.addAttribute("pv", pv);
+		model.addAttribute("searchVo", searchVo);
 
 		return "board/free";
 	}
@@ -68,25 +68,27 @@ public class BoardController {
 	@GetMapping("board/market")
 	public String getMarketList(@RequestParam(defaultValue="1") String p, @RequestParam Map<String, String> searchVo, Model model, HttpSession session) {
 		
-		try {
-			
-			searchVo.put("cno", "2");
-
-			int listCount = service.countBoard(searchVo);
-			int currentPage = Integer.parseInt(p);
-			int pageLimit = 5;
-			int boardLimit = 8;
-			PageVo pv = new PageVo(listCount, currentPage, pageLimit, boardLimit);
-		
-			List<BoardVo> voList = service.getList(pv, searchVo);
-			
-			model.addAttribute("voList", voList);
-			model.addAttribute("pv", pv);
-			model.addAttribute("searchVo", searchVo);
-			
-		}catch (Exception e) {
-			e.printStackTrace();
+		//관리자도, 회원도 아니면 리턴
+		AdminVo loginAdmin = (AdminVo) session.getAttribute("loginAdmin");
+		MemberVo loginMember = (MemberVo) session.getAttribute("loginMember");
+		if(loginAdmin == null && loginMember == null) {
+			session.setAttribute("alertMsg", "로그인 후 이용해주세요.");
+			return "redirect:/member/mlogin";
 		}
+			
+		searchVo.put("cno", "2");
+
+		int listCount = service.countBoard(searchVo);
+		int currentPage = Integer.parseInt(p);
+		int pageLimit = 5;
+		int boardLimit = 8;
+		PageVo pv = new PageVo(listCount, currentPage, pageLimit, boardLimit);
+	
+		List<BoardVo> voList = service.getList(pv, searchVo);
+		
+		model.addAttribute("voList", voList);
+		model.addAttribute("pv", pv);
+		model.addAttribute("searchVo", searchVo);
 
 		return "board/market";
 	}
@@ -95,25 +97,27 @@ public class BoardController {
 	@GetMapping("board/noname")
 	public String getNonameList(@RequestParam(defaultValue="1") String p, @RequestParam Map<String, String> searchVo, Model model, HttpSession session) {
 		
-		try {
-			
-			searchVo.put("cno", "3");
-
-			int listCount = service.countBoard(searchVo);
-			int currentPage = Integer.parseInt(p);
-			int pageLimit = 5;
-			int boardLimit = 8;
-			PageVo pv = new PageVo(listCount, currentPage, pageLimit, boardLimit);
-		
-			List<BoardVo> voList = service.getList(pv, searchVo);
-			
-			model.addAttribute("voList", voList);
-			model.addAttribute("pv", pv);
-			model.addAttribute("searchVo", searchVo);
-			
-		}catch (Exception e) {
-			e.printStackTrace();
+		//관리자도, 회원도 아니면 리턴
+		AdminVo loginAdmin = (AdminVo) session.getAttribute("loginAdmin");
+		MemberVo loginMember = (MemberVo) session.getAttribute("loginMember");
+		if(loginAdmin == null && loginMember == null) {
+			session.setAttribute("alertMsg", "로그인 후 이용해주세요.");
+			return "redirect:/member/mlogin";
 		}
+			
+		searchVo.put("cno", "3");
+
+		int listCount = service.countBoard(searchVo);
+		int currentPage = Integer.parseInt(p);
+		int pageLimit = 5;
+		int boardLimit = 8;
+		PageVo pv = new PageVo(listCount, currentPage, pageLimit, boardLimit);
+	
+		List<BoardVo> voList = service.getList(pv, searchVo);
+		
+		model.addAttribute("voList", voList);
+		model.addAttribute("pv", pv);
+		model.addAttribute("searchVo", searchVo);
 
 		return "board/noname";
 	}
@@ -122,25 +126,27 @@ public class BoardController {
 	@GetMapping("board/praise")
 	public String getPraiseList(@RequestParam(defaultValue="1") String p, @RequestParam Map<String, String> searchVo, Model model, HttpSession session) {
 		
-		try {
-			
-			searchVo.put("cno", "4");
-
-			int listCount = service.countBoard(searchVo);
-			int currentPage = Integer.parseInt(p);
-			int pageLimit = 5;
-			int boardLimit = 8;
-			PageVo pv = new PageVo(listCount, currentPage, pageLimit, boardLimit);
-		
-			List<BoardVo> voList = service.getList(pv, searchVo);
-			
-			model.addAttribute("voList", voList);
-			model.addAttribute("pv", pv);
-			model.addAttribute("searchVo", searchVo);
-			
-		}catch (Exception e) {
-			e.printStackTrace();
+		//관리자도, 회원도 아니면 리턴
+		AdminVo loginAdmin = (AdminVo) session.getAttribute("loginAdmin");
+		MemberVo loginMember = (MemberVo) session.getAttribute("loginMember");
+		if(loginAdmin == null && loginMember == null) {
+			session.setAttribute("alertMsg", "로그인 후 이용해주세요.");
+			return "redirect:/member/mlogin";
 		}
+			
+		searchVo.put("cno", "4");
+
+		int listCount = service.countBoard(searchVo);
+		int currentPage = Integer.parseInt(p);
+		int pageLimit = 5;
+		int boardLimit = 8;
+		PageVo pv = new PageVo(listCount, currentPage, pageLimit, boardLimit);
+	
+		List<BoardVo> voList = service.getList(pv, searchVo);
+		
+		model.addAttribute("voList", voList);
+		model.addAttribute("pv", pv);
+		model.addAttribute("searchVo", searchVo);
 
 		return "board/praise";
 	}
@@ -148,7 +154,14 @@ public class BoardController {
 	
 	// 게시글 작성 (화면)
 	@GetMapping("board/write")
-	public String write(Model model) {
+	public String write(HttpSession session, Model model) {
+		//회원 아니면 리턴
+		MemberVo loginMember = (MemberVo) session.getAttribute("loginMember");
+		if(loginMember == null) {
+			session.setAttribute("alertMsg", "회원으로 로그인 해주세요.");
+			return "redirect:/board/free";
+		}
+		
 		List<BoardCategoryVo> cvo = service.getCategory();
 		model.addAttribute("cvo", cvo);
 		return "board/write";
@@ -157,43 +170,36 @@ public class BoardController {
 	// 게시글 작성
 	@PostMapping("board/write")
 	public String write(HttpSession session, BoardVo vo, String imgList) {
-				
-		try {
-			
-//			AdminVo loginAdmin = (AdminVo) session.getAttribute("loginAdmin");
-//			if(loginAdmin == null) {
-//				return "redirect:/"
-//			}
-//			vo.setWriterNo(loginAdmin.getNo());
-			
-			vo.setWriterNo("2"); // 임시 작성자번호
-			int result = service.write(vo);
-			if(result != 1) {
-				session.setAttribute("alert", "게시글 작성 실패...");
-				return "redirect:/board/free";
-			}
-			
-			
-			//썸네일 db 저장
-			System.out.println(imgList);
-			BoardImgVo ivo = new BoardImgVo();
-			
-			if (imgList != null && !"".equals(imgList)) {
-				String[] arr = imgList.split(",");
-				ivo.setName(arr[0]);
-				int imgResult = service.insertImgToDb(ivo);
-				System.out.println("이미지 저장 결과 : " + imgResult);
-			}
-					
-//			for (String s : arr) {
-//				System.out.println("배열에 담은 사진 : " + s);
-//				ivo.setName(s);
-//				service.insertImgToDb(ivo);
-//			}
-			
-		}catch (Exception e) {
-			e.printStackTrace();
+		
+		//회원 아니면 리턴
+		MemberVo loginMember = (MemberVo) session.getAttribute("loginMember");
+		if(loginMember == null) {
+			session.setAttribute("alertMsg", "회원으로 로그인 해주세요.");
+			return "redirect:/board/free";
 		}
+				
+		vo.setWriterNo(loginMember.getNo());
+		
+		int result = service.write(vo);
+		if(result != 1) {
+			session.setAttribute("alert", "게시글 작성 실패...");
+			return "redirect:/board/free";
+		}
+		
+		
+		//썸네일 db 저장
+		BoardImgVo ivo = new BoardImgVo();
+		if (imgList != null && !"".equals(imgList)) {
+			String[] arr = imgList.split(",");
+			ivo.setName(arr[0]);
+			service.insertImgToDb(ivo);
+		}
+				
+//		for (String s : arr) {
+//			System.out.println("배열에 담은 사진 : " + s);
+//			ivo.setName(s);
+//			service.insertImgToDb(ivo);
+//		}
 		
 		session.setAttribute("alert", "게시글 작성 성공!");
 		
@@ -236,79 +242,102 @@ public class BoardController {
     @GetMapping("board/edit")
     public String edit(HttpSession session, Model model, String no) {
     	
-		List<BoardCategoryVo> cvo = service.getCategory();
+		//no에 부적절한 값 입력시 리턴
+	    try {
+	        int intNo = Integer.parseInt(no);
+	        if (intNo <= 0) {
+	            session.setAttribute("alertMsg", "잘못된 글번호입니다.");
+	            return "redirect:/board/free";
+	        }
+	    } catch (NumberFormatException e) {
+	        session.setAttribute("alertMsg", "잘못된 글번호입니다.");
+	        return "redirect:/board/free";
+	    }
+    	
 		BoardVo vo = service.viewDetail(no);
-		model.addAttribute("cvo", cvo);
 		model.addAttribute("vo", vo);
-		if(no == null || "".equals(no) || "0".equals(no)) {
-			session.setAttribute("alert", "잘못된 접근입니다.");
+    	
+		//회원이 아니거나, 작성자가 아닐 때
+		MemberVo loginMember = (MemberVo) session.getAttribute("loginMember");
+		if(loginMember == null || ! vo.getWriterNo().equals( loginMember.getNo() ) ) {
+			session.setAttribute("alertMsg", "작성한 계정으로 로그인 해주세요.");
 			return "redirect:/board/free";
 		}
+		
+		
+		List<BoardCategoryVo> cvo = service.getCategory();
+		model.addAttribute("cvo", cvo);
     	
     	return "board/edit";
     }
+    
     
 	//게시글 수정
 	@PostMapping("board/edit")
 	public String edit(HttpSession session, BoardVo vo) {
 		
-		try {
-			
-			vo.setWriterNo("2");
-			int result = service.edit(vo);
-			if(result != 1) {
-				session.setAttribute("alertMsg", "게시글 수정 실패...");
-				return "redirect:/board/detail?no=" + vo.getNo();
-			}
-		}catch (Exception e) {
-			e.printStackTrace();
+		//회원이 아니거나, 작성자가 아닐 때
+		MemberVo loginMember = (MemberVo) session.getAttribute("loginMember");
+		vo.setWriterNo(loginMember.getNo());
+		if(loginMember == null || ! vo.getWriterNo().equals( loginMember.getNo() ) ) {
+			session.setAttribute("alertMsg", "작성한 계정으로 로그인 해주세요.");
+			return "redirect:/board/free";
+		}
+		
+		int result = service.edit(vo);
+		
+		if(result != 1) {
+			session.setAttribute("alertMsg", "게시글 수정 실패...");
+			return "redirect:/board/detail?no=" + vo.getNo();
 		}
 		
 		session.setAttribute("alertMsg", "게시글 수정 성공!");
 		return "redirect:/board/detail?no=" + vo.getNo();
 	}
 	
+	
 	//게시글 삭제
 	@GetMapping("board/delete")
 	public String delete(HttpSession session, String no) {
 		
-		//리턴값 카테고리 위해 vo 생성
-		BoardVo vo = service.viewDetail(no);
+		//no에 부적절한 값 입력시 리턴
+	    try {
+	        int intNo = Integer.parseInt(no);
+	        if (intNo <= 0) {
+	            session.setAttribute("alertMsg", "잘못된 글번호입니다.");
+	            return "redirect:/board/free";
+	        }
+	    } catch (NumberFormatException e) {
+	        session.setAttribute("alertMsg", "잘못된 글번호입니다.");
+	        return "redirect:/board/free";
+	    }
 		
-		try {
-			
-			AdminVo loginAdmin = (AdminVo) session.getAttribute("loginAdmin");
-//			if (loginAdmin == null) {
-//				session.setAttribute("alertMsg", "게시글 삭제 실패...");
-//				return "redirect:/main/amain";
-//			}
-			
-			//String writerNo = loginAdmin.getNo();
-			String writerNo = "2";
-			
-			Map<String, String> noMap = new HashMap<>();
-			noMap.put("writerNo", writerNo);
-			noMap.put("no", no);
-			
-			int result = service.delete(noMap);
+		BoardVo vo = service.viewDetail(no);
+		//작성자이거나, 관리소장이면 삭제
+		MemberVo loginMember = (MemberVo) session.getAttribute("loginMember");
+		AdminVo loginAdmin = (AdminVo) session.getAttribute("loginAdmin");
+		if( vo.getWriterNo().equals(loginMember.getNo()) || "M".equals(loginAdmin.getGrade()) ) {
+
+			int result = service.delete(no);
 			if(result != 1) {
 				session.setAttribute("alertMsg", "게시글 삭제 실패...");
 				return "redirect:/board/detail?no=" + no;
 			}
+				
+			session.setAttribute("alertMsg", "게시글 삭제 성공 !");
 			
-		}catch (Exception e) {
-			e.printStackTrace();
-		}
-		
-		session.setAttribute("alertMsg", "게시글 삭제 성공 !");
-		
-		if ("2".equals( vo.getBoardCno() )) {
-			return "redirect:/board/market";
-		}else if ("3".equals( vo.getBoardCno() )) {
-			return "redirect:/board/noname";
-		}else if ("4".equals( vo.getBoardCno() )) {
-			return "redirect:/board/praise";
+			if ("2".equals( vo.getBoardCno() )) {
+				return "redirect:/board/market";
+			}else if ("3".equals( vo.getBoardCno() )) {
+				return "redirect:/board/noname";
+			}else if ("4".equals( vo.getBoardCno() )) {
+				return "redirect:/board/praise";
+			}else {
+				return "redirect:/board/free";
+			}
+			
 		}else {
+			session.setAttribute("alertMsg", "작성한 계정으로 로그인 해주세요.");
 			return "redirect:/board/free";
 		}
 		
@@ -317,15 +346,36 @@ public class BoardController {
 	
 	// 게시글 상세 조회
 	@GetMapping("board/detail")
-	public String viewDetail(@RequestParam(defaultValue="1")String no, Model model) {
+	public String viewDetail(@RequestParam(defaultValue="1")String no, HttpSession session, Model model) {
 		
-		try {
-			BoardVo vo = service.viewDetail(no);
-			model.addAttribute("vo", vo);
-		}catch (Exception e) {
-			e.printStackTrace();
+		//관리자도, 회원도 아니면 리턴
+		AdminVo loginAdmin = (AdminVo) session.getAttribute("loginAdmin");
+		MemberVo loginMember = (MemberVo) session.getAttribute("loginMember");
+		if(loginAdmin == null && loginMember == null) {
+			session.setAttribute("alertMsg", "로그인 후 이용해주세요.");
+			return "redirect:/member/mlogin";
 		}
 		
+		//no에 부적절한 값 입력시 리턴
+	    try {
+	        int intNo = Integer.parseInt(no);
+	        if (intNo <= 0) {
+	            session.setAttribute("alertMsg", "잘못된 글번호입니다.");
+	            return "redirect:/board/free";
+	        }
+	    } catch (NumberFormatException e) {
+	        session.setAttribute("alertMsg", "잘못된 글번호입니다.");
+	        return "redirect:/board/free";
+	    }
+		
+		
+		BoardVo vo = service.viewDetail(no);
+		if (vo == null) {
+			session.setAttribute("alertMsg", "없는 게시글입니다.");
+			return "redirect:/board/free";
+		}
+		model.addAttribute("vo", vo);
+			
 		return "board/detail";
 	}
 	
@@ -335,7 +385,12 @@ public class BoardController {
 	@ResponseBody
 	public String writeReply(HttpSession session, ReplyVo rvo) {
 		
-		AdminVo loginMember = (AdminVo) session.getAttribute("loginMember");
+		//회원이 아니면 실패
+		MemberVo loginMember = (MemberVo) session.getAttribute("loginMember");
+		if(loginMember == null) {
+			session.setAttribute("alertMsg", "회원으로 로그인 해주세요.");
+			return "fail";
+		}
 		
 		//댓글 null 검사
 		if (rvo.getContent() == null || rvo.getContent().trim().isEmpty()) {
@@ -357,7 +412,12 @@ public class BoardController {
 	@ResponseBody
 	public String editReply(HttpSession session, ReplyVo rvo) {
 		
-		AdminVo loginMember = (AdminVo) session.getAttribute("loginMember");
+		//회원이 아니거나, 작성자가 아닐 때
+		MemberVo loginMember = (MemberVo) session.getAttribute("loginMember");
+		if( (loginMember!=null && !loginMember.getNo().equals(rvo.getWriterNo())) || loginMember == null ) {
+			session.setAttribute("alertMsg", "작성한 계정으로 로그인 해주세요.");
+			return "fail";
+		}
 		
 		//댓글 null 검사
 		if (rvo.getContent() == null || rvo.getContent().trim().isEmpty()) {
@@ -377,22 +437,22 @@ public class BoardController {
 	//댓글 삭제
 	@PostMapping("reply/delete")
 	@ResponseBody
-	public String deleteReply(HttpSession session, ReplyVo rvo) {
+	public String deleteReply(HttpSession session, ReplyVo rvo, String adminGrade) {
 		
-		try {
-			
-			AdminVo loginMember = (AdminVo) session.getAttribute("loginMember");
-			
+		//작성자이거나, 관리소장이면 삭제
+		MemberVo loginMember = (MemberVo) session.getAttribute("loginMember");
+		
+		if( (loginMember != null && rvo.getWriterNo().equals(loginMember.getNo())) || ("M".equals(adminGrade) && loginMember == null) ) {
+
 			int result = service.deleteReply(rvo);
-			
 			if(result <= 0) {
 				return "fail";
 			}
-			
-		}catch (Exception e) {
-			e.printStackTrace();
+
+		}else {
+			return "fail";
 		}
-		
+			
 		return "success";
 		
 	}
@@ -402,13 +462,7 @@ public class BoardController {
 	@ResponseBody
 	public List<ReplyVo> getReplyList(HttpSession session, String boardNo) throws JsonProcessingException {
 		
-		MemberVo loginMember = (MemberVo) session.getAttribute("loginMember");
-		
 		List<ReplyVo> rvoList = service.getReplyList(boardNo);
-		
-//		if(rvoList == null) {
-//			return "fail";
-//		}
 		
 		//자바객체 -> JSON 형태
 //		ObjectMapper mapper = new ObjectMapper();
@@ -419,9 +473,9 @@ public class BoardController {
 	}
 	
 	//좋아요
-	@GetMapping("love")
+	@PostMapping("love")
 	@ResponseBody
-	public int clickLove(LoveVo lvo) {
+	public int clickLove(HttpSession session, LoveVo lvo) {
 		
 		//좋아요 여부
 		int loveYn = service.checkLoveYn(lvo);
@@ -439,6 +493,5 @@ public class BoardController {
 		return loveCnt;
 		
 	}
-	
 
 }
